+++ source/Public/Get-NetworkScan.ps1
@@ -13,7 +13,7 @@ function Get-NetworkScan {
         Get-NetworkScan -report
     .PARAMETER Ports
         Default ports are:
-        "21", "22", "25", "53", "67", "68", "80", "443", `
+        "21", "22", "23", "25", "53", "67", "68", "80", "443", `
         "88", "464", "123", "135", "137", "138", "139", `
         "445", "389", "636", "514", "587", "1701", `
         "3268", "3269", "3389", "5985", "5986"
@@ -52,11 +52,12 @@ function Get-NetworkScan {
     begin {
         If (Get-Module -ListAvailable -Name "PSnmap") { Import-Module "PSnmap" } Else { Install-Module "PSnmap" -Force; Import-Module "PSnmap" }
         if (!($ports)) {
-            [int[]]$ports = "21", "22", "25", "53", "67", "68", "80", "443", `
+            [int[]]$ports = "21", "22", "23", "25", "53", "67", "68", "80", "443", `
                 "88", "464", "123", "135", "137", "138", "139", `
                 "445", "389", "636", "514", "587", "1701", `
                 "3268", "3269", "3389", "5985", "5986"
         }
+        $ouiobject = Invoke-RestMethod https://standards-oui.ieee.org/oui/oui.csv | ConvertFrom-Csv
     } # Begin Close
     process {
         if ($LocalSubnets) {
@@ -73,11 +74,25 @@ function Get-NetworkScan {
                     $NetWorkScan = Invoke-PSnmap -ComputerName $subnet -Port $ports -Dns -NoSummary -AddService
                     # Filter devices that don't ping as no results will be found.
                     $scan = $NetworkScan | Where-Object { $_.Ping -eq $true }
-                    Write-Output "##########################################"
-                    Write-Output "Network scan for Subnet $Subnet completed."
-                    Write-Output "DHCP Server: $($DHCPServer)"
-                    Write-Output "Gateway: $($network.IPv4DefaultGateway.nexthop)"
-                    Write-Output "##########################################"
+                    Write-Verbose "##########################################"
+                    Write-Verbose "Network scan for Subnet $Subnet completed."
+                    Write-Verbose "DHCP Server: $($DHCPServer)"
+                    Write-Verbose "Gateway: $($network.IPv4DefaultGateway.nexthop)"
+                    Write-Verbose "##########################################"
+                    $scan | ForEach-Object {
+                        $org = ""
+                        $macid = ((arp -a $_.ComputerName | Select-String '([0-9a-f]{2}-){5}[0-9a-f]{2}').Matches.Value).Replace("-", ":")
+                        $macpop = $macid.replace(":", "")
+                        $macsubstr = $macpop.Substring(0, 6)
+                        $org = ($ouiobject | Where-Object { $_.assignment -eq $macsubstr })."Organization Name"
+                        Add-Member -InputObject $_ -MemberType NoteProperty -Name MacID -Value $macid
+                        if ($org) {
+                            Add-Member -InputObject $_ -MemberType NoteProperty -Name ManufacturerName -Value $org
+                        }
+                        else {
+                            Add-Member -InputObject $_ -MemberType NoteProperty -Name ManufacturerName -Value "Not Found"
+                        }
+                    }
                     # Normalize Subnet text for filename.
                     $subnetText = $(($subnet.Replace("/", ".CIDR.")))
                     # If report switch is true.